Extract recommendation region helpers
